+++ client/src/javascript/components/torrent-list/TorrentList.js
@@ -3,11 +3,11 @@ import _ from 'lodash';
 import Dropzone from 'react-dropzone';
 import React from 'react';
 
-import ContextMenu from '../general/ContextMenu';
 import ConfigStore from '../../stores/ConfigStore';
 import CustomScrollbars from '../general/CustomScrollbars';
 import EventTypes from '../../constants/EventTypes';
 import Files from '../icons/Files';
+import GlobalContextMenuMountPoint from '../general/GlobalContextMenuMountPoint';
 import ListViewport from '../general/ListViewport';
 import LoadingIndicator from '../general/LoadingIndicator';
 import PriorityMeter from '../general/filesystem/PriorityMeter';
@@ -185,7 +185,7 @@ class TorrentListContainer extends React.Component {
   }
 
   getContextMenuItems(torrent) {
-    let clickHandler = this.handleContextMenuItemClick;
+    const clickHandler = this.handleContextMenuItemClick;
 
     return [{
       action: 'start',
@@ -633,7 +633,7 @@ class TorrentListContainer extends React.Component {
           ref={ref => this.horizontalScrollRef = ref}>
           <div className="torrent__list__wrapper"
             style={listWrapperStyle}>
-            <ContextMenu id="torrent-list-item" />
+            <GlobalContextMenuMountPoint id="torrent-list-item" />
             {torrentListHeading}
             {content}
           </div>
